+++ de.haumacher.msgbuf.generator/src/test/java/test/openworld/novisitor/base/impl/Shape_Impl.java
@@ -1,0 +1,223 @@
+package test.openworld.novisitor.base.impl;
+
+/**
+ * Implementation of {@link test.openworld.novisitor.base.Shape}.
+ */
+public abstract class Shape_Impl extends de.haumacher.msgbuf.data.AbstractDataObject implements test.openworld.novisitor.base.Shape {
+
+	static {
+		de.haumacher.msgbuf.data.TypeRegistryLoader.ensureLoaded();
+	}
+
+	private String _name = "";
+
+	/**
+	 * Creates a {@link Shape_Impl} instance.
+	 */
+	public Shape_Impl() {
+		super();
+	}
+
+	@Override
+	public final String getName() {
+		return _name;
+	}
+
+	@Override
+	public test.openworld.novisitor.base.Shape setName(String value) {
+		internalSetName(value);
+		return this;
+	}
+
+	/** Internal setter for {@link #getName()} without chain call utility. */
+	protected final void internalSetName(String value) {
+		_listener.beforeSet(this, NAME__PROP, value);
+		_name = value;
+		_listener.afterChanged(this, NAME__PROP);
+	}
+
+	protected de.haumacher.msgbuf.observer.Listener _listener = de.haumacher.msgbuf.observer.Listener.NONE;
+
+	@Override
+	public test.openworld.novisitor.base.Shape registerListener(de.haumacher.msgbuf.observer.Listener l) {
+		internalRegisterListener(l);
+		return this;
+	}
+
+	protected final void internalRegisterListener(de.haumacher.msgbuf.observer.Listener l) {
+		_listener = de.haumacher.msgbuf.observer.Listener.register(_listener, l);
+	}
+
+	@Override
+	public test.openworld.novisitor.base.Shape unregisterListener(de.haumacher.msgbuf.observer.Listener l) {
+		internalUnregisterListener(l);
+		return this;
+	}
+
+	protected final void internalUnregisterListener(de.haumacher.msgbuf.observer.Listener l) {
+		_listener = de.haumacher.msgbuf.observer.Listener.unregister(_listener, l);
+	}
+
+	protected static final java.util.List<String> PROPERTIES;
+	static {
+		java.util.List<String> local = java.util.Arrays.asList(
+			NAME__PROP);
+		PROPERTIES = java.util.Collections.unmodifiableList(local);
+	}
+
+	protected static final java.util.Set<String> TRANSIENT_PROPERTIES;
+	static {
+		java.util.HashSet<String> tmp = new java.util.HashSet<>();
+		tmp.addAll(java.util.Arrays.asList(
+				));
+		TRANSIENT_PROPERTIES = java.util.Collections.unmodifiableSet(tmp);
+	}
+
+	@Override
+	public java.util.List<String> properties() {
+		return PROPERTIES;
+	}
+
+	@Override
+	public java.util.Set<String> transientProperties() {
+		return TRANSIENT_PROPERTIES;
+	}
+
+	@Override
+	public Object get(String field) {
+		switch (field) {
+			case NAME__PROP: return getName();
+			default: return test.openworld.novisitor.base.Shape.super.get(field);
+		}
+	}
+
+	@Override
+	public void set(String field, Object value) {
+		switch (field) {
+			case NAME__PROP: internalSetName((String) value); break;
+		}
+	}
+
+	@Override
+	public final void writeTo(de.haumacher.msgbuf.json.JsonWriter out) throws java.io.IOException {
+		out.beginArray();
+		out.value(jsonType());
+		writeContent(out);
+		out.endArray();
+	}
+
+	@Override
+	protected void writeFields(de.haumacher.msgbuf.json.JsonWriter out) throws java.io.IOException {
+		super.writeFields(out);
+		out.name(NAME__PROP);
+		out.value(getName());
+	}
+
+	@Override
+	protected void readField(de.haumacher.msgbuf.json.JsonReader in, String field) throws java.io.IOException {
+		switch (field) {
+			case NAME__PROP: setName(de.haumacher.msgbuf.json.JsonUtil.nextStringOptional(in)); break;
+			default: super.readField(in, field);
+		}
+	}
+
+	/** XML element name representing a {@link test.openworld.novisitor.base.Shape} type. */
+	public static final String SHAPE__XML_ELEMENT = "shape";
+
+	/** XML attribute or element name of a {@link #getName} property. */
+	private static final String NAME__XML_ATTR = "name";
+
+	@Override
+	public final void writeContent(javax.xml.stream.XMLStreamWriter out) throws javax.xml.stream.XMLStreamException {
+		writeAttributes(out);
+		writeElements(out);
+	}
+
+	/** Serializes all fields that are written as XML attributes. */
+	protected void writeAttributes(javax.xml.stream.XMLStreamWriter out) throws javax.xml.stream.XMLStreamException {
+		out.writeAttribute(NAME__XML_ATTR, getName());
+	}
+
+	/** Serializes all fields that are written as XML elements. */
+	protected void writeElements(javax.xml.stream.XMLStreamWriter out) throws javax.xml.stream.XMLStreamException {
+		// No element fields.
+	}
+
+	/** Creates a new {@link test.openworld.novisitor.base.Shape} and reads properties from the content (attributes and inner tags) of the currently open element in the given {@link javax.xml.stream.XMLStreamReader}. */
+	public static Shape_Impl readShape_XmlContent(javax.xml.stream.XMLStreamReader in) throws javax.xml.stream.XMLStreamException {
+		switch (in.getLocalName()) {
+			case Circle_Impl.CIRCLE__XML_ELEMENT: {
+				return test.openworld.novisitor.base.impl.Circle_Impl.readCircle_XmlContent(in);
+			}
+
+			default: {
+				de.haumacher.msgbuf.data.Factory<? extends test.openworld.novisitor.base.Shape> factory = test.openworld.novisitor.base.Shape.XML_REGISTRY.get(in.getLocalName());
+				test.openworld.novisitor.base.Shape instance = factory == null ? null : factory.create();
+				if (instance instanceof Shape_Impl) {
+					Shape_Impl result = (Shape_Impl) instance;
+					result.readContentXml(in);
+					return result;
+				}
+				internalSkipUntilMatchingEndElement(in);
+				return null;
+			}
+		}
+	}
+
+	/** Reads properties from the content (attributes and inner tags) of the currently open element in the given {@link javax.xml.stream.XMLStreamReader}. */
+	protected final void readContentXml(javax.xml.stream.XMLStreamReader in) throws javax.xml.stream.XMLStreamException {
+		for (int n = 0, cnt = in.getAttributeCount(); n < cnt; n++) {
+			String name = in.getAttributeLocalName(n);
+			String value = in.getAttributeValue(n);
+
+			readFieldXmlAttribute(name, value);
+		}
+		while (true) {
+			int event = in.nextTag();
+			if (event == javax.xml.stream.XMLStreamConstants.END_ELEMENT) {
+				break;
+			}
+			assert event == javax.xml.stream.XMLStreamConstants.START_ELEMENT;
+
+			String localName = in.getLocalName();
+			readFieldXmlElement(in, localName);
+		}
+	}
+
+	/** Parses the given attribute value and assigns it to the field with the given name. */
+	protected void readFieldXmlAttribute(String name, String value) {
+		switch (name) {
+			case NAME__XML_ATTR: {
+				setName(value);
+				break;
+			}
+			default: {
+				// Skip unknown attribute.
+			}
+		}
+	}
+
+	/** Reads the element under the cursor and assigns its contents to the field with the given name. */
+	protected void readFieldXmlElement(javax.xml.stream.XMLStreamReader in, String localName) throws javax.xml.stream.XMLStreamException {
+		switch (localName) {
+			case NAME__XML_ATTR: {
+				setName(in.getElementText());
+				break;
+			}
+			default: {
+				internalSkipUntilMatchingEndElement(in);
+			}
+		}
+	}
+
+	protected static final void internalSkipUntilMatchingEndElement(javax.xml.stream.XMLStreamReader in) throws javax.xml.stream.XMLStreamException {
+		int level = 0;
+		while (true) {
+			switch (in.next()) {
+				case javax.xml.stream.XMLStreamConstants.START_ELEMENT: level++; break;
+				case javax.xml.stream.XMLStreamConstants.END_ELEMENT: if (level == 0) { return; } else { level--; break; }
+			}
+		}
+	}
+
+}